Make resources thread-safe

+++ src/Layers/xrRender/ResourceManager.cpp
@@ -460,7 +460,7 @@ void	CResourceManager::_DumpMemoryUsage		()
 		{
 			u32			m = I->second->flags.MemoryUsage;
 			shared_str	n = I->second->cName;
-			mtex.insert (std::make_pair(m,std::make_pair(I->second->dwReference,n) ));
+			mtex.insert (std::make_pair(m,std::make_pair((u32)I->second->dwReference,n) ));
 		}
 	}
